+++ src/corpus/corpus.composable.js
@@ -5,8 +5,6 @@ import useMessenger from "@/message/messenger.composable";
 import useCorpora from "@/corpora/corpora.composable";
 import useConfig from "./config/config.composable";
 import useExports from "./exports/exports.composable";
-import useSources from "./sources/sources.composable";
-import useJob from "./job/job.composable";
 
 /** Let data be refreshed initially, but skip subsequent load calls. */
 const isCorpusFresh = {};
@@ -19,21 +17,31 @@ export default function useCorpus(corpusId) {
   const { alertError } = useMessenger();
   const { loadConfig } = useConfig(corpusId);
   const { loadExports } = useExports(corpusId);
-  const { loadJob } = useJob(corpusId);
-  const { loadSources } = useSources(corpusId);
 
   async function loadCorpus(force = false) {
+    // Make sure the corpus has an entry in the store.
     await loadCorpora();
     if (isCorpusFresh[corpusId] && !force) {
       return;
     }
+
+    // Load all essential info about the corpus.
     await Promise.all([
       loadConfig(), //
       loadExports(),
-      loadJob(),
-      loadSources(),
+      loadResourceInfo(),
     ]);
+
+    // Remember to skip loading next time.
     isCorpusFresh[corpusId] = true;
+  }
+
+  /** Load job status and source files in the same request. */
+  async function loadResourceInfo() {
+    const info = await mink.resourceInfo(corpusId).catch(alertError);
+    corpusStore.corpora[corpusId].name = info.resource.name;
+    corpusStore.corpora[corpusId].sources = info.resource.source_files;
+    corpusStore.corpora[corpusId].status = info.job;
   }
 
   async function deleteCorpus(corpusId_ = corpusId) {